refactor: changes in responses and types

+++ src/CrowdParlay.Users.Infrastructure.Persistence/Services/UsersRepository.cs
@@ -1,3 +1,4 @@
+using CrowdParlay.Users.Application.Exceptions;
 using CrowdParlay.Users.Domain.Abstractions;
 using CrowdParlay.Users.Domain.Entities;
 using CrowdParlay.Users.Infrastructure.Persistence.Abstractions;
@@ -79,6 +80,14 @@ internal class UsersRepository : IUsersRepository
     public async Task DeleteAsync(Uuid id, CancellationToken cancellationToken = default)
     {
         await using var connection = await _connectionFactory.CreateConnectionAsync(cancellationToken);
+        var isUserExists = await connection.QueryAsync(
+            $"SELECT COUNT(*) FROM {UserSchema.Table} WHERE {UserSchema.Id} = @{nameof(id)}",
+            new { id }  
+        );
+
+        if (!isUserExists.Any())
+            throw new NotFoundException();
+        
         await connection.ExecuteAsync(
             $@"DELETE FROM {UserSchema.Table} WHERE {UserSchema.Id} = @{nameof(id)}",
             new { id });
